improve example

+++ modules/example_rust/src/lib.rs
@@ -1,10 +1,10 @@
 #[cxx::bridge]
 mod ffi {
     extern "Rust" {
-        fn add_rs(left: isize, right: isize) -> isize;
+        fn add_rs(left: i32, right: i32) -> i32;
     }
 }
 
-pub fn add_rs(left: isize, right: isize) -> isize {
+pub fn add_rs(left: i32, right: i32) -> i32 {
     left + right
 }
